# golang-backend/docs/swagger.yaml
basePath: /
definitions:
  controller.BaseSwaggerResp:
    properties:
      code:
        example: 0
        type: integer
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
  controller.SwaggerChangePasswordReq:
    properties:
      confirmPassword:
        type: string
      currentPassword:
        type: string
      newPassword:
        type: string
      newUsername:
        type: string
    required:
    - confirmPassword
    - currentPassword
    - newPassword
    - newUsername
    type: object
...
  controller.SwaggerLoginReq:
    properties:
      captchaData: {}
      captchaId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  controller.SwaggerLoginResp:
    properties:
      code:
        example: 0
        type: integer
      data:
        properties:
          name:
            example: admin
            type: string
          requirePasswordChange:
            example: false
            type: boolean
          role_id:
            example: 0
            type: integer
          token:
            example: Bearer xxx
            type: string
        type: object
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
...
  controller.SwaggerRegisterReq:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: demo
        type: string
    type: object
  controller.SwaggerRegisterResp:
    properties:
      code:
        example: 0
        type: integer
      data:
        properties:
          name:
            example: admin
            type: string
          requirePasswordChange:
            example: false
            type: boolean
          role_id:
            example: 0
            type: integer
          token:
            example: Bearer xxx
            type: string
        type: object
      msg:
        example: 操作成功
        type: string
      ts:
        example: 1700000000000
        type: integer
    type: object
  controller.SwaggerResetFlowReq:
    properties:
      id:
        type: integer
      type:
        description: 1 user, else tunnel
        type: integer
    required:
    - id
    - type
    type: object
...
  controller.SwaggerUserCreateReq:
    properties:
      expTime:
        type: integer
      flow:
        type: integer
      flowResetTime:
        type: integer
      num:
        type: integer
      pwd:
        type: string
      status:
        type: integer
      user:
        type: string
    required:
    - pwd
    - user
    type: object
...
  controller.SwaggerUserUpdateReq:
    properties:
      expTime:
        type: integer
      flow:
        type: integer
      flowResetTime:
        type: integer
      id:
        type: integer
      num:
        type: integer
      pwd:
        type: string
      status:
        type: integer
      user:
        type: string
    required:
    - id
    type: object
...
info:
  contact: {}
  description: 面板后端接口文档
  title: network-panel API
  version: "1.0"
paths:
...
  /api/v1/user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerUserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 创建用户
      tags:
      - user
  /api/v1/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 需要提供 id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerUserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 删除用户
      tags:
      - user
...
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: 登录并返回 token
      parameters:
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerLoginResp'
      summary: 用户登录
      tags:
      - user
...
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: 公开注册，受配置 registration_enabled 控制
      parameters:
      - description: 注册参数（用户名、密码）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SwaggerRegisterResp'
      summary: 用户注册
      tags:
      - user
  /api/v1/user/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: type=1 用户流量，否则隧道路由
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerResetFlowReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 重置用户/隧道流量
      tags:
      - user
  /api/v1/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户更新参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerUserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 更新用户
      tags:
      - user
  /api/v1/user/updatePassword:
    post:
      consumes:
      - application/json
      parameters:
      - description: 修改密码参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controller.SwaggerChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BaseSwaggerResp'
      summary: 修改当前用户密码
      tags:
      - user
...
swagger: "2.0"